+++ s/base/index.ts
@@ -1,4 +1,5 @@
 
+export * from "./utils/states.js"
 export * from "./css-reset.js"
 export * from "./element.js"
 export * from "./use.js"